+++ CHANGELOG.md
@@ -1518,3 +1518,30 @@ _2023-09-10_
 
 - 修复`TnNotify`修改颜色失败的问题
 - 修复`TnTabs`在支付宝小程序端出错的问题
+
+### 1.0.17-beta.1
+
+#### Fixed
+
+- 修复`TnTabs`设置`active-font-size`无效的问题
+- 修复`TnPicker`不选择后直接点击确认会报错的问题
+
+### 1.0.17-beta.2
+
+#### 新增
+
+- `TnButton`新增`getRealtimePhoneNumber`type类型
+- `TnButton`新增`phoneNumberNoQuotaToast`参数
+- `TnButton`新增`getrealtimephonenumber`事件
+
+### 1.0.17-beta.4
+
+#### 新增
+
+- `TnButton`新增`contact`事件
+
+### 1.0.17-beta.5
+
+#### Fixed
+
+- 修复`TnCountDown`无法动态修改倒计时的问题